program: port lookup table status into processor
